things

+++ libs/genetic-algorithm/src/chromosome.rs
@@ -1,7 +1,7 @@
 use std::ops::Index;
 
 pub struct Chromosome {
-    genes: Vec<f64>,
+    genes: Vec<f32>,
 }
 
 #[allow(clippy::len_without_is_empty)]
@@ -10,25 +10,25 @@ impl Chromosome {
         self.genes.len()
     }
 
-    pub fn iter(&self) -> impl Iterator<Item = &f64> {
+    pub fn iter(&self) -> impl Iterator<Item = &f32> {
         self.genes.iter()
     }
 
-    pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut f64> {
+    pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut f32> {
         self.genes.iter_mut()
     }
 }
 
 impl Index<usize> for Chromosome {
-    type Output = f64;
+    type Output = f32;
 
     fn index(&self, index: usize) -> &Self::Output {
         &self.genes[index]
     }
 }
 
-impl FromIterator<f64> for Chromosome {
-    fn from_iter<T: IntoIterator<Item = f64>>(iter: T) -> Self {
+impl FromIterator<f32> for Chromosome {
+    fn from_iter<T: IntoIterator<Item = f32>>(iter: T) -> Self {
         Self {
             genes: iter.into_iter().collect(),
         }
@@ -36,8 +36,8 @@ impl FromIterator<f64> for Chromosome {
 }
 
 impl IntoIterator for Chromosome {
-    type Item = f64;
-    type IntoIter = impl Iterator<Item = f64>;
+    type Item = f32;
+    type IntoIter = impl Iterator<Item = f32>;
 
     fn into_iter(self) -> Self::IntoIter {
         self.genes.into_iter()
