+++ libs/simulation/src/animal_individual.rs
@@ -1,12 +1,12 @@
 use crate::*;
 
 pub struct AnimalIndividual {
-    fitness: f64,
+    fitness: f32,
     chromosome: ga::Chromosome,
 }
 
 impl ga::Individual for AnimalIndividual {
-    fn fitness(&self) -> f64 {
+    fn fitness(&self) -> f32 {
         self.fitness
     }
 
@@ -25,7 +25,7 @@ impl ga::Individual for AnimalIndividual {
 impl AnimalIndividual {
     pub fn from_animal(animal: &Animal) -> Self {
         Self {
-            fitness: animal.collisions as f64,
+            fitness: animal.collisions as f32,
             chromosome: animal.as_chromosome(),
         }
     }
